requirements.txt
pysdl2
desper<2
corso